+++ omitdcm-compiler-plugin/src/main/kotlin/com/cakcaraka/omitdcm/compiler/OmitDCMIrGenerationExtension.kt
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.cakcaraka.omitdcm.compiler
 
+import com.cakcaraka.omitdcm.annotations.OmitPropertyStrategy
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
@@ -9,6 +10,7 @@ import org.jetbrains.kotlin.name.ClassId
 
 public class OmitDCMIrGenerationExtension(
   private val omitToStringAnnotations: Set<ClassId>,
+  private val defaultOmitPropertyStrategy: OmitPropertyStrategy,
 ) : IrGenerationExtension {
 
   override fun generate(moduleFragment: IrModuleFragment, pluginContext: IrPluginContext) {
@@ -16,6 +18,7 @@ public class OmitDCMIrGenerationExtension(
       OmitDCMIrVisitor(
         pluginContext,
         omitToStringAnnotations,
+        defaultOmitPropertyStrategy,
       )
     moduleFragment.transform(transformer, null)
   }