+++ CHANGELOG.md
@@ -6,6 +6,6 @@ Changelog
 
 - Initial release.
 - Create skeleton based on [ZacSweers/redacted-compiler-plugin](https://github.com/ZacSweers/redacted-compiler-plugin).
-- Support `@OmitToString` annotation to omit data class `toString()` implementations.
+- Support `@OmitDCMToString` annotation to omit data class `toString()` implementations.
 - Kotlin JVM only.
 - Publish to JitPack.